+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - mame: state save/restore ("-state save/restore") supported via MAME's 'statesave'/'stateload' debugger commands.
 - Fixed setting negative values for memory elements of 4.
 - Memory views are now accessible directly from the "Variable" and "Watch" pane.
+- Data watchpoints can be set directly from the "Watch" pane.
 
 # 3.7.4
 - PR #193: Fixed bug where pressing 'stop' would sometimes take 5 seconds to disconnect. (mame)